% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.p}
\alias{.p}
\title{Data warehouse path}
\usage{
.p(
  parquet_root = Sys.getenv(x = "PARQUET_ROOT", unset = system.file("data/warehouse",
    package = "HRDashboard"))
)
}
\description{
Data warehouse path
}
\keyword{internal}
