% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_choices.R
\name{.parquet}
\alias{.parquet}
\title{Parquet glob path}
\usage{
.parquet(
  parquet_root = Sys.getenv(x = "PARQUET_ROOT", unset = system.file("data/warehouse",
    package = "HRDashboard"))
)
}
\description{
Parquet glob path
}
\keyword{internal}
