% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{.pp}
\alias{.pp}
\title{Payroll data warehouse path}
\usage{
.pp(
  parquet_root = Sys.getenv(x = "PARQUET_ROOT", unset = system.file("data/warehouse",
    package = "HRDashboard"))
)
}
\description{
Payroll data warehouse path
}
\keyword{internal}
