% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_conn.R
\name{duckdb_conn}
\alias{duckdb_conn}
\title{Create a DuckDB connection}
\usage{
duckdb_conn(
  duckdb_path = Sys.getenv(x = "DUCKDB_PATH", unset = file.path(system.file(package =
    "HRDashboard"), "data", "base", "attendance.duckdb"))
)
}
\arguments{
\item{duckdb_path}{String. Path to the DuckDB database file.
Default is read from the environment variable "DUCKDB_PATH".}
}
\value{
DBI::dbConnect()
}
\description{
Create a DuckDB connection
}
