% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_choices.R
\name{get_departments_choices}
\alias{get_departments_choices}
\title{Get departments choices for selectize inputs}
\usage{
get_departments_choices()
}
\value{
Named list suitable for \code{selectizeInput}.
}
\description{
Get departments choices for selectize inputs
}
