% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_choices.R
\name{get_employees_choices}
\alias{get_employees_choices}
\title{Get employees choices for selectize inputs}
\usage{
get_employees_choices(dept_id = "all")
}
\arguments{
\item{dept_id}{Integer vector or \code{"all"}. Filter by department(s).}
}
\value{
Named list (optgroup style) suitable for \code{selectizeInput}.
}
\description{
Get employees choices for selectize inputs
}
