% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_choices.R
\name{get_locations_choices}
\alias{get_locations_choices}
\title{Get office/location choices for selectize inputs}
\usage{
get_locations_choices()
}
\value{
Named character vector suitable for \code{selectInput}.
}
\description{
Get office/location choices for selectize inputs
}
