% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric_choices.R
\name{get_metric_parameters}
\alias{get_metric_parameters}
\title{Get metric parameters}
\usage{
get_metric_parameters(category, metric)
}
\arguments{
\item{category}{String. Category name (see \code{get_category_choices}).}

\item{metric}{String. Metric id within the category.}
}
\value{
Named list with \code{id}, \code{label}, \code{suffix}.
}
\description{
Get metric parameters
}
