% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_choices.R
\name{get_years_choices}
\alias{get_years_choices}
\title{Get year choices for selectize inputs}
\usage{
get_years_choices()
}
\value{
Named character vector suitable for \code{selectInput}.
}
\description{
Get year choices for selectize inputs
}
