% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_absenteeism_rate}
\alias{query_absenteeism_rate}
\title{Absenteeism rate from attendance Parquet}
\usage{
query_absenteeism_rate(
  date_from = NULL,
  date_to = NULL,
  dept_ids = NULL,
  loc_ids = NULL
)
}
\description{
Absenteeism rate from attendance Parquet
}
