% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_age_distribution}
\alias{query_age_distribution}
\title{Age distribution (10-year buckets)}
\usage{
query_age_distribution(dept_ids = NULL, loc_ids = NULL)
}
\description{
Age distribution (10-year buckets)
}
