% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_education_distribution}
\alias{query_education_distribution}
\title{Education level distribution}
\usage{
query_education_distribution(dept_ids = NULL, loc_ids = NULL)
}
\description{
Education level distribution
}
