% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_gender_diversity}
\alias{query_gender_diversity}
\title{Gender diversity breakdown}
\usage{
query_gender_diversity(
  by = c("department", "location", "level"),
  dept_ids = NULL,
  loc_ids = NULL
)
}
\arguments{
\item{by}{One of "department", "location", "level".}
}
\description{
Gender diversity breakdown
}
