% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_hours_distribution}
\alias{query_hours_distribution}
\title{Hours-worked distribution in 0.5-hour buckets}
\usage{
query_hours_distribution(
  date_from = NULL,
  date_to = NULL,
  dept_ids = NULL,
  emp_ids = NULL,
  loc_ids = NULL
)
}
\arguments{
\item{date_from, date_to}{Date range.}

\item{dept_ids, emp_ids, loc_ids}{Optional filter vectors.}
}
\description{
Hours-worked distribution in 0.5-hour buckets
}
