% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_payroll_cost_trend}
\alias{query_payroll_cost_trend}
\title{Payroll trend (monthly) from payroll Parquet}
\usage{
query_payroll_cost_trend(dept_ids = NULL, loc_ids = NULL, year = NULL)
}
\description{
Payroll trend (monthly) from payroll Parquet
}
