% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_payroll_kpi_summary}
\alias{query_payroll_kpi_summary}
\title{Monthly payroll KPI summary (across employees)}
\usage{
query_payroll_kpi_summary(
  year = NULL,
  month = NULL,
  dept_ids = NULL,
  loc_ids = NULL
)
}
\description{
Monthly payroll KPI summary (across employees)
}
