% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query_tenure_distribution}
\alias{query_tenure_distribution}
\title{Tenure distribution (years)}
\usage{
query_tenure_distribution(dept_ids = NULL, loc_ids = NULL)
}
\description{
Tenure distribution (years)
}
