% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_last_checktime.R
\name{read_last_checktime}
\alias{read_last_checktime}
\title{Check last checktime (refresh time)}
\usage{
read_last_checktime(
  checktime_file = Sys.getenv(x = "CHECKTIME_FILE", unset =
    system.file("data/state/last_checktime.txt", package = "HRDashboard"))
)
}
\description{
Check last checktime (refresh time)
}
