% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_live_attendance.R
\name{read_live_attendance}
\alias{read_live_attendance}
\title{Read live attendance for today from DuckDB}
\usage{
read_live_attendance()
}
\value{
data.table with columns: employee_id, name, dept_name,
check_type, check_time, diff_mins, trend.
}
\description{
Read live attendance for today from DuckDB
}
